バトル開始とバトル終了の処理を実装し、一応バトルと言えるまでにはなりました

+++ Assets/Datas/Parameters/Bonus/BonusBase.cs
@@ -4,7 +4,7 @@ namespace Parameter{
 	public abstract class BonusBase {
 		protected string name;
 		protected int bonusValue;
-		protected int limit;
+		protected float limit;
 
 
 		//このボーナスの名前を返します
